switched to std::size_t for matrix(-matrix) operations

+++ src/dataStructures/RightHandSide.cpp
@@ -9,5 +9,5 @@ RightHandSide::RightHandSide(const std::size_t N, const int blockSize, sycl::que
     blockCountX(std::ceil(static_cast<double>(N) / static_cast<double>(blockSize))),
     rightHandSideData(sycl::usm_allocator<conf::fp_type, sycl::usm::alloc::shared>(queue)) {
     // allocate memory for right-hand side storage
-    rightHandSideData.resize(blockCountX * blockSize);
+    rightHandSideData.resize(static_cast<std::size_t>(blockCountX) * static_cast<std::size_t>(blockSize));
 }
